+++ src/api/index.ts
@@ -1,7 +1,5 @@
 export * from './endpoints/auth.api';
 export * from './endpoints/location.api';
 export * from './endpoints/trip.api';
-
 export * from './services/apiRoutes.constants';
-
 export { default as API } from './services/axios';