+++ src/api/services/apiRoutes.constants.ts
@@ -4,8 +4,10 @@ export const apiRoutes = {
     location: 'location',
     trip: 'trip',
     tripSearch: 'trip/search',
+    allBookings: 'booking',
     booking: 'booking',
     bookingPnr: 'booking/pnr',
     userBooking: 'booking/user/all',
     user: 'user',
+    myBookings: 'booking/user/all',
 };