+++ src/api/types/ticket.ts
@@ -1,4 +1,5 @@
 import { ITicket } from '../../types/ticket';
+import { IUserExternal } from './user';
 import { ITripExternal } from './trip';
 
 export interface ITicketRequest {