+++ src/components/DatagridListingpage/datagrid/CustomPagination/CustomPagination.styled.ts
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+import Box from '@mui/material/Box';
+
+export const CustomPaginationWrapper = styled(Box)(() => ({
+    display: 'flex',
+    alignItems: 'center',
+    justifyContent: 'space-between',
+    padding: '1.6rem',
+    width: '100%',
+}));