+++ src/config/constants.ts
@@ -5,10 +5,12 @@ export const paths = {
     home: '/',
     tripsListing: '/trips',
     tripDetail: '/trips/:tripId',
+    bookings: '/bookings',
     usersListing: '/users',
     tripBooking: '/trips/booking',
     // ticket: '/ticket/:pnrNumber',
     ticket: '/ticket',
+    myBookings: '/user/bookings',
 };
 
 export const colors = {
@@ -22,13 +24,14 @@ export const colors = {
     bookedSeat: '#bdbdbd',
     darkBookedSeat: '#8a8a8a',
     seat: '#fcfcfc',
-    darkSeat: '#d2d2d2',
-    busLayoutBg: '#eeeeee',
+    darkSeat: '#e1dede',
+    busLayoutBg: '#E8DEF8',
     darkBusLayoutBg: '#575757',
     red: '#D84E55',
     boxShadowPrimary: 'rgba(0,0,0,0.26)',
     green: '#56A55C',
     redHover: '#AB3840',
+    linkColor: '#0000EE',
 };
 
 // api format: id, name => mui format: id, label