+++ src/config/routes/routes.config.tsx
@@ -3,7 +3,9 @@ import { LandingPage, TripsListingPage, TripBookingPage } from '../../pages';
 import { ErrorPage } from '../../pages/ErrorPage/ErrorPage';
 import { Layout } from '../../layout';
 import { paths } from '..';
+import { AllBookingsPage } from '../../pages/BookingsList/AllBookingsPage/AllBookingsPage';
 import { UsersListingPage } from '../../pages/UserListing/UsersListingPage';
+import { MyBookingsPage } from '../../pages/BookingsList/MyBookingsPage/MyBookingsPage';
 import { RequireAuth } from '../../components/RequireAuth/RequireAuth';
 import { EUserRole } from '../../types';
 import { TicketPage } from '../../pages/TicketPage/TicketPage';
@@ -28,6 +30,10 @@ export const routesConfig: RouteObject[] = [
                         path: paths.usersListing,
                         element: <UsersListingPage />,
                     },
+                    {
+                        path: paths.bookings,
+                        element: <AllBookingsPage />,
+                    },
                 ],
             },
             {
@@ -50,6 +56,10 @@ export const routesConfig: RouteObject[] = [
             {
                 path: paths.tripBooking,
                 element: <TripBookingPage />,
+            },
+            {
+                path: paths.myBookings,
+                element: <MyBookingsPage />,
             },
         ],
     },